Fixes #3670 by using separate routers for sends and publishes.
